Move random util to root level test/utils
