+++ framework/test/mocha/functional/http/get/node/transactions_pending.js
@@ -22,7 +22,7 @@ const {
 	registerMultisignature,
 } = require('@liskhq/lisk-transactions');
 const apiHelpers = require('../../../../common/helpers/api');
-const randomUtil = require('../../../../common/utils/random');
+const randomUtil = require('../../../../../utils/random');
 const SwaggerEndpoint = require('../../../../common/swagger_spec');
 const accountFixtures = require('../../../../fixtures/accounts');
 const waitFor = require('../../../../common/utils/wait_for');
