+++ framework/test/mocha/functional/http/post/0.transfer.js
@@ -23,7 +23,7 @@ const typesRepresentatives = require('../../../fixtures/types_representatives');
 const phases = require('../../../common/phases');
 const sendTransactionPromise = require('../../../common/helpers/api')
 	.sendTransactionPromise;
-const randomUtil = require('../../../common/utils/random');
+const randomUtil = require('../../../../utils/random');
 const apiCodes = require('../../../../../src/modules/http_api/api_codes');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
