+++ framework/test/mocha/functional/http/post/4.multisignature.js
@@ -19,7 +19,7 @@ const { registerMultisignature } = require('@liskhq/lisk-transactions');
 const phases = require('../../../common/phases');
 const Scenarios = require('../../../common/scenarios');
 const accountFixtures = require('../../../fixtures/accounts');
-const randomUtil = require('../../../common/utils/random');
+const randomUtil = require('../../../../utils/random');
 const waitFor = require('../../../common/utils/wait_for');
 const elements = require('../../../common/utils/elements');
 const SwaggerEndpoint = require('../../../common/swagger_spec');
