+++ framework/test/mocha/functional/http/post/6.dapps.in_transfer.js
@@ -20,7 +20,7 @@ const BigNum = require('@liskhq/bignum');
 const Promise = require('bluebird');
 const phases = require('../../../common/phases');
 const accountFixtures = require('../../../fixtures/accounts');
-const randomUtil = require('../../../common/utils/random');
+const randomUtil = require('../../../../utils/random');
 const waitFor = require('../../../common/utils/wait_for');
 const apiHelpers = require('../../../common/helpers/api');
 const apiCodes = require('../../../../../src/modules/http_api/api_codes');
