+++ framework/test/mocha/functional/ws/transport/transport.js
@@ -20,7 +20,7 @@ const { transfer } = require('@liskhq/lisk-transactions');
 const { generatePeerHeader } = require('../../../common/generatePeerHeader');
 const waitFor = require('../../../common/utils/wait_for');
 const SwaggerEndpoint = require('../../../common/swagger_spec');
-const randomUtil = require('../../../common/utils/random');
+const randomUtil = require('../../../../utils/random');
 const accountFixtures = require('../../../fixtures/accounts');
 const apiHelpers = require('../../../common/helpers/api');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
