+++ framework/test/mocha/integration/blocks/chain/apply_block.js
@@ -19,7 +19,7 @@ const expect = require('chai').expect;
 const { transfer, registerDelegate } = require('@liskhq/lisk-transactions');
 
 const accountFixtures = require('../../../fixtures/accounts');
-const randomUtil = require('../../../common/utils/random');
+const randomUtil = require('../../../../utils/random');
 const localCommon = require('../../common');
 const blocksChainModule = require('../../../../../src/modules/chain/blocks/chain');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
