+++ framework/test/mocha/integration/get/transactions_unconfirmed.js
@@ -17,7 +17,7 @@
 const async = require('async');
 const { transfer } = require('@liskhq/lisk-transactions');
 const accountFixtures = require('../../fixtures/accounts');
-const randomUtil = require('../../common/utils/random');
+const randomUtil = require('../../../utils/random');
 const localCommon = require('./../common');
 const { getNetworkIdentifier } = require('../../common/network_identifier');
