+++ framework/test/mocha/integration/rebuilding.js
@@ -16,7 +16,7 @@
 
 const Promise = require('bluebird');
 const { transfer } = require('@liskhq/lisk-transactions');
-const randomUtil = require('../common/utils/random');
+const randomUtil = require('../../utils/random');
 const accountsFixtures = require('../fixtures/accounts');
 const QueriesHelper = require('../common/integration/sql/queries_helper');
 const localCommon = require('./common');
