+++ framework/test/mocha/integration/rounds.js
@@ -28,7 +28,7 @@ const Promise = require('bluebird');
 const { hexToBuffer } = require('@liskhq/lisk-cryptography');
 const { Slots } = require('../../../src/modules/chain/dpos');
 const accountsFixtures = require('../fixtures/accounts');
-const randomUtil = require('../common/utils/random');
+const randomUtil = require('../../utils/random');
 const QueriesHelper = require('../common/integration/sql/queries_helper');
 const localCommon = require('./common');
 const { getNetworkIdentifier } = require('../common/network_identifier');
