+++ framework/test/mocha/integration/transactions/0_0_transfer.js
@@ -16,7 +16,7 @@
 
 const { transfer } = require('@liskhq/lisk-transactions');
 const accountFixtures = require('../../fixtures/accounts');
-const randomUtil = require('../../common/utils/random');
+const randomUtil = require('../../../utils/random');
 const localCommon = require('../common');
 const { getNetworkIdentifier } = require('../../common/network_identifier');
