+++ framework/test/mocha/integration/transactions/1_second_signature/1_1_second_signature.js
@@ -19,7 +19,7 @@ const {
 	registerSecondPassphrase,
 } = require('@liskhq/lisk-transactions');
 const accountFixtures = require('../../../fixtures/accounts');
-const randomUtil = require('../../../common/utils/random');
+const randomUtil = require('../../../../utils/random');
 const localCommon = require('../../common');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
