+++ framework/test/mocha/integration/transactions/2_delegates/1_same_account_same_username.js
@@ -16,7 +16,7 @@
 
 const { transfer, registerDelegate } = require('@liskhq/lisk-transactions');
 const accountFixtures = require('../../../fixtures/accounts');
-const randomUtil = require('../../../common/utils/random');
+const randomUtil = require('../../../../utils/random');
 const localCommon = require('../../common');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
