+++ framework/test/mocha/integration/transactions/2_delegates/2_same_account_different_usernames.js
@@ -16,7 +16,7 @@
 
 const { transfer, registerDelegate } = require('@liskhq/lisk-transactions');
 const accountFixtures = require('../../../fixtures/accounts');
-const randomUtil = require('../../../common/utils/random');
+const randomUtil = require('../../../../utils/random');
 const localCommon = require('../../common');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
