+++ framework/test/mocha/integration/transactions/2_delegates/5_same_account_from_pool_and_peer.js
@@ -18,7 +18,7 @@ const { transfer, registerDelegate } = require('@liskhq/lisk-transactions');
 const expect = require('chai').expect;
 const accountFixtures = require('../../../fixtures/accounts');
 const localCommon = require('../../common');
-const randomUtil = require('../../../common/utils/random');
+const randomUtil = require('../../../../utils/random');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
 
 const networkIdentifier = getNetworkIdentifier(
