+++ framework/test/mocha/integration/transactions/3_3_votes.js
@@ -16,7 +16,7 @@
 
 const { transfer, castVotes } = require('@liskhq/lisk-transactions');
 const accountFixtures = require('../../fixtures/accounts');
-const randomUtil = require('../../common/utils/random');
+const randomUtil = require('../../../utils/random');
 const localCommon = require('../common');
 const { getNetworkIdentifier } = require('../../common/network_identifier');
