+++ framework/test/mocha/integration/transactions/4_multisignature/4_X_multisignature_edge_cases.js
@@ -21,7 +21,7 @@ const {
 	createSignatureObject,
 } = require('@liskhq/lisk-transactions');
 const accountFixtures = require('../../../fixtures/accounts');
-const randomUtil = require('../../../common/utils/random');
+const randomUtil = require('../../../../utils/random');
 const localCommon = require('../../common');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
