+++ framework/test/mocha/integration/transactions/4_multisignature/4_X_multisignature_validated.js
@@ -19,7 +19,7 @@ const {
 	registerMultisignature,
 	createSignatureObject,
 } = require('@liskhq/lisk-transactions');
-const randomUtil = require('../../../common/utils/random');
+const randomUtil = require('../../../../utils/random');
 const Scenarios = require('../../../common/scenarios');
 const localCommon = require('../../common');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
