+++ framework/test/mocha/integration/transactions/4_multisignature/duplicate_signatures.js
@@ -22,7 +22,7 @@ const {
 	createSignatureObject,
 } = require('@liskhq/lisk-transactions');
 const accountsFixtures = require('../../../fixtures/accounts');
-const randomUtil = require('../../../common/utils/random');
+const randomUtil = require('../../../../utils/random');
 const localCommon = require('../../common');
 const { getNetworkIdentifier } = require('../../../common/network_identifier');
