+++ framework/test/mocha/integration/transactions/inert_transactions.js
@@ -7,7 +7,7 @@ const {
 } = require('@liskhq/lisk-transactions');
 const localCommon = require('../common');
 const accountFixtures = require('../../fixtures/accounts');
-const randomUtil = require('../../common/utils/random');
+const randomUtil = require('../../../utils/random');
 const { getNetworkIdentifier } = require('../../common/network_identifier');
 
 const networkIdentifier = getNetworkIdentifier(
